re-organize files in ruleng project
