+++ src/aggregator-ruleng/Engine/RuleExecutionContext.cs
