+++ src/aggregator-ruleng/Engine/RuleExecutor.cs
