+++ src/aggregator-ruleng/Engine/ScriptedRuleWrapper.cs
