+++ src/aggregator-ruleng/Engine/WorkItemData.cs
