+++ src/aggregator-ruleng/Extensions/WorkItemExtension.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Collections.ObjectModel;
+using Microsoft.TeamFoundation.WorkItemTracking.WebApi.Models;
+using Microsoft.VisualStudio.Services.Common;
+
+
+namespace aggregator.Engine
+{
+    public static class WorkItemExtension
+    {
+        public static string GetTeamProject(this WorkItem workItem)
+        {
+            return workItem.Fields.GetCastedValueOrDefault("System.TeamProject", default(string));
+        }
+    }
+}
