+++ src/aggregator-ruleng/Interfaces/IAggregatorLogger.cs
