+++ src/aggregator-ruleng/Interfaces/IRule.cs
