+++ src/aggregator-ruleng/Interfaces/IRuleProvider.cs
