+++ src/aggregator-ruleng/Interfaces/IRuleResult.cs
