+++ src/aggregator-ruleng/Interfaces/IRuleSettings.cs
