+++ src/aggregator-ruleng/Parsing/IPreprocessedRule.cs
