+++ src/aggregator-ruleng/Parsing/PreprocessedRule.cs
