+++ src/aggregator-ruleng/Parsing/PreprocessedRuleExtensions.cs
