+++ src/aggregator-ruleng/Parsing/RuleSettings.cs
