+++ src/aggregator-ruleng/Persistance/JsonPatchOperationConverter.cs
