+++ src/aggregator-ruleng/Persistance/RelationPatch.cs
