+++ src/aggregator-ruleng/RuleObjects/WorkItemId.cs
