+++ src/aggregator-ruleng/RuleObjects/WorkItemStore.cs
