+++ src/aggregator-ruleng/RuleObjects/WorkItemUpdateWrapper.cs
