+++ src/aggregator-ruleng/RuleObjects/WorkItemWrapper.cs
